requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
google-generativeai
elevenlabs==1.4.0
pydub